Pydantic 2 upgrade (#57)
This Pr just ups the version identifier for pydantic in pyproject.toml
from `1.*` to `2.*`. the `bump-pydantic` tool didn't find anything to
change, and running skellycam with pydantic 2.7 seemed to work properly.

## Extra Testing:
There are pydantic references in the timestamp plotting code, which
doesn't run on mac. Someone with a windows or linux machine should run
through a recording to make sure the plots still generate correctly.

---------

Co-authored-by: jonmatthis <jonmatthis@gmail.com>