Accept different charuco boards (#77)
This PR allows the new charuco board selection option to be used with
the charuco overlay option. This means the 5x3 board will be able to be
detected for the overlay option, not just the original 7x5 board.

It works by creating a function for `SkellyCamWidget` to accept
different charuco board names, that it then passes to the
`CamGroupThreadWorker`, which creates a charuco board object that is
used for the Charuco Overlay option.

Right now, I'm passing the name of the board into skellycam. We can pass
the entire board definition in, which will require changing freemocap's
board definition to include a detector object (adding this to the
__post_init__: `self.charuco_detector =
cv2.aruco.CharucoDetector(self.charuco_board)`)